Double values are now allowed in otu tables and are not just cast as integers
git-svn-id: svn+ssh://svn.code.sf.net/p/topiaryexplorer/code/trunk/main@537 05d1262b-b726-4fff-8b9f-3609a8a82e43

+++ src/topiaryexplorer/TreeWindow.java
@@ -804,16 +804,16 @@ public class TreeWindow extends TopiaryWindow implements KeyListener, ActionList
                          Object w = row.get(i); // get the OTU count
                          //if it's not a Number, skip it
                          if (!(w instanceof Number)) continue;
-                         Integer weight = 0;
+                         double weight = 0;
                          try {
-                             weight = (Integer)w;
-                         } catch ( ClassCastException e) {
-                             weight = ((Double)w).intValue();
+                             weight = (Double)w;
+                         } catch (ClassCastException e) {
+                             weight = ((Integer)w).doubleValue();
                          }
-                         if (weight == 0) continue; // this sample doesn't contain this OTU
+                         if (weight == 0.0) continue; // this sample doesn't contain this OTU
                          
                          if(!weighted)
-                            weight = 1;
+                            weight = 1.0;
                          
                          String sampleID = frame.otuSampleMap.getColumnName(((Number)i).intValue());
                          // System.out.println("sampleID:"+sampleID);
